Add type restriction for audio uploading. File upload box and related structure has been changed.
