+++ listening-test-react/src/index.tsx
@@ -5,20 +5,10 @@ import * as serviceWorker from './serviceWorker';
 import App from "./App";
 import 'mobx-react/batchingForReactDom'
 import Axios from "axios";
-import {createMuiTheme} from "@material-ui/core";
-import {red} from "@material-ui/core/colors";
 
 Axios.defaults.withCredentials = true;
 Axios.defaults.xsrfHeaderName = "X-CSRFToken";
 Axios.defaults.xsrfCookieName = "_xsrf";
-
-const theme = createMuiTheme({
-  palette: {
-    warning: {
-      main: red[500],
-    },
-  },
-});
 
 ReactDOM.render(
   <React.StrictMode>
