+++ listening-test-react/src/views/audio/AbTest/AbTestItemExampleCard.tsx
@@ -8,7 +8,7 @@ import AudioExampleSettingsDialog from "../AudioExampleSettingsDialog";
 import {observer} from "mobx-react";
 import {SurveyControlType} from "../../../shared/models/EnumsAndTypes";
 import {TestItemExampleCardProps} from "../../components/TypesAndItemOverrides";
-import {TestItemCardFileDropGrid} from "../../components/TestItemCardFileDropGrid";
+import {AudioFileDropGrid} from "../AudioFileDropGrid";
 import {RemovableSurveyControl} from "../../../shared/components/RemovableSurveyControl";
 
 export const AbTestItemExampleCard = observer((props: React.PropsWithChildren<TestItemExampleCardProps>) => {
@@ -30,7 +30,7 @@ export const AbTestItemExampleCard = observer((props: React.PropsWithChildren<Te
           <Grid item xs={12}>
             <TagsGroup value={example.tags} onChange={newTags => example.tags = newTags}/>
           </Grid>
-          <TestItemCardFileDropGrid example={example} reference keepPlace/>
+          <AudioFileDropGrid example={example} reference keepPlace/>
           {/*Special survey questions for ab test*/}
           {example.fields?.map((q, qi) => <Grid item xs={12} key={qi}>
             <RemovableSurveyControl question={q} hideRemoveButton={qi < 1}
