+++ listening-test-react/src/views/audio/AcrTest/AcrTestItemExampleCard.tsx
@@ -1,16 +1,17 @@
 import Card from "@material-ui/core/Card";
 import CardHeader from "@material-ui/core/CardHeader";
-import {CardContent, Collapse} from "@material-ui/core";
+import {CardContent, Collapse, Typography} from "@material-ui/core";
 import {SurveyControl} from "../../../shared/components/SurveyControl";
 import React from "react";
 import {TagsGroup} from "../../../shared/components/TagsGroup";
 import Grid from "@material-ui/core/Grid";
-import {FileDropZone} from "../../../shared/components/FileDropZone";
 import AudioExampleSettingsDialog from "../AudioExampleSettingsDialog";
 import {observer} from "mobx-react";
 import {TestItemExampleCardProps} from "../../components/TypesAndItemOverrides";
-import {TestItemCardFileDropGrid} from "../../components/TestItemCardFileDropGrid";
-import {AudioFileModel, AudioExampleSettingsModel} from "../../../shared/models/AudioTestModel";
+import {AudioFileDropGrid} from "../AudioFileDropGrid";
+import {AudioExampleSettingsModel, AudioFileModel} from "../../../shared/models/AudioTestModel";
+import Icon from "@material-ui/core/Icon";
+import {FileUploadDropBox} from "../../../shared/components/FileUploadDropBox";
 
 export const AcrTestItemExampleCard = observer((props: React.PropsWithChildren<TestItemExampleCardProps>) => {
   const {example, title, action, collapsed} = props;
@@ -35,11 +36,14 @@ export const AcrTestItemExampleCard = observer((props: React.PropsWithChildren<T
             <SurveyControl control={q}/>
           </Grid>)}
 
-          <TestItemCardFileDropGrid example={example} reference/>
+          <AudioFileDropGrid example={example} reference/>
 
           {/*Placeholder for adding to list*/}
           <Grid item xs={12} md={4}>
-            <FileDropZone onChange={handleAdd} label="Drop or click to add a file"/>
+            <FileUploadDropBox onChange={handleAdd} fileType="audio">
+              <Typography>Drop or click to add a file</Typography>
+              <Icon>attachment</Icon>
+            </FileUploadDropBox>
           </Grid>
         </Grid>
       </CardContent>
