+++ listening-test-react/src/views/audio/AudioTestItemTrainingCard.tsx
@@ -6,14 +6,15 @@ import {SurveyControlModel} from "../../shared/models/SurveyControlModel";
 import Card from "@material-ui/core/Card";
 import CardHeader from "@material-ui/core/CardHeader";
 import AudioExampleSettingsDialog from "./AudioExampleSettingsDialog";
-import {CardContent, Collapse} from "@material-ui/core";
+import {CardContent, Collapse, Typography} from "@material-ui/core";
 import Grid from "@material-ui/core/Grid";
-import {TestItemCardFileDropGrid} from "../components/TestItemCardFileDropGrid";
+import {AudioFileDropGrid} from "./AudioFileDropGrid";
 import {SurveyControl} from "../../shared/components/SurveyControl";
-import {FileDropZone} from "../../shared/components/FileDropZone";
 import {RemovableSurveyControl} from "../../shared/components/RemovableSurveyControl";
 import {useMatStyles} from "../SharedStyles";
 import {AddQuestionButton} from "../../shared/components/AddQuestionButton";
+import Icon from "@material-ui/core/Icon";
+import {FileUploadDropBox} from "../../shared/components/FileUploadDropBox";
 
 /** Training Test Item: Audios will play synchronously*/
 export const AudioTestItemTrainingCard = observer((props: React.PropsWithChildren<TestItemExampleCardProps>) => {
@@ -38,10 +39,13 @@ export const AudioTestItemTrainingCard = observer((props: React.PropsWithChildre
           {example.fields && example.fields[0] && <Grid item xs={12}>
             <SurveyControl control={example.fields[0]}/>
           </Grid>}
-          <TestItemCardFileDropGrid example={example}/>
+          <AudioFileDropGrid example={example}/>
           {/*Placeholder for adding to list*/}
           <Grid item xs={12} md={4}>
-            <FileDropZone onChange={handleAdd} label="Drop or click to add a file"/>
+            <FileUploadDropBox onChange={handleAdd} fileType="audio">
+              <Typography>Drop or click to add a file</Typography>
+              <Icon>attachment</Icon>
+            </FileUploadDropBox>
           </Grid>
           {/*Additional question add*/}
           {example.fields && !(example.fields.length > 1) ? <Grid item xs={12} className={classes.flexEnd}>
